requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
openai>=1.55.0
pyodbc==5.0.1
python-multipart==0.0.6
requests==2.31.0
psycopg2-binary==2.9.9
PyJWT>=2.8.0
gunicorn==21.2.0
